backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
numpy==1.24.3
scipy==1.10.1
networkx==3.1
google-generativeai==0.3.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
scikit-image==0.21.0
scikit-learn==1.3.2
moviepy==1.0.3
librosa==0.10.1
Pillow==10.0.1
